NO JIRA. Removed checkConnection for daemon.
CheckConnection checks of Dataverse and easy-validate-dans-bag can both be reached and failed they cannot.
On system reboot if one of these services is slower to start than dd-dans-deposit-to-dataverse this results in
dd-dans-deposit-to-dataverse not starting.

This behavior should be changed to dd-dans-deposit polling the services it depends on for a longer period before
giving up and failing to start.

For now not checking at all is the least bad option.